Added GetLAPSPassword to examples. (#1)
* Added GetLAPSPassword to examples.

* update GetLapsPassword

* Add LAPSv2 column

---------
